fixing bianry search

+++ rtrlib/aspa/aspa_array/aspa_array.c
@@ -145,13 +145,13 @@ size_t aspa_array_search(struct aspa_array *vector, uint32_t customer_asn)
 	}
 
 	// left and right bound of our search space
-	size_t left = 0;
-	size_t right = vector->size;
+	long left = 0;
+	long right = vector->size;
 
 	// we stop if right and left crossed
 	while (left <= right) {
 		// current center
-		size_t center = left + ((right - left) / 2);
+		long center = left + ((right - left) / 2);
 		uint32_t center_value = vector->data[center].customer_asn;
 
 		if (center_value == customer_asn) {
